+++ arki/system.py
@@ -1,16 +1,27 @@
+import os
 import sys
 
 
 def is_linux():
+    """
+    Check if the current platform is Linux.
+    :return: True if the current platform is Linux; False otherwise
+    """
     return sys.platform.startswith("linux")
 
 
 def is_windows():
+    """
+    Check if the current platform is Windows.
+    :return: True if the current platform is Windows; False otherwise
+    """
     return sys.platform.startswith("win")
 
 
 def print_export_env(env_dict):
-    """Export environment variables of the given dictionary
+    """
+    Print the command(s) for exporting environment variables to current platform
+    :param env_dict: Dictionary containing the environment variables
     """
     if is_linux():
         cmd_env_str = "export {}={}"
@@ -25,3 +36,15 @@ def print_export_env(env_dict):
         print(cmd_env_str.format(k, v))
     print("-----------------------------------------")
 
+
+def run_command(command):
+    """
+    Run a command
+    :param command: command string
+    :return: return code (0 means all good)
+    """
+    print(command)
+    ret_code = os.system(command)
+
+    print(f"Command return code: {ret_code}")
+    return ret_code