+++ ReleaseNotes.md
@@ -3,4 +3,5 @@
 
 - **`arki`**: Show all tools
 - **`aws_profile`**: Look up AWS profile and access key and print the export commands to console.
-- **`env_store`**: Support saving and retrieving environment variables.
+- **`env_store`**: Support saving and retrieving some environment variables in a local file.
+- **`venv`**: Create `virtualenv` of a Python version on the current platform (Linux or Windows).